+++ plugins/org.eclipse.n4js.xpect/src/org/eclipse/n4js/xpect/common/N4JSXpectRunner.java
@@ -49,8 +49,7 @@ public class N4JSXpectRunner extends XpectRunner {
 			}
 		}
 
-		IXpectURIProvider provider = new LspCompatibleXpectTestFiles.LspCompatibleXpectTestFileCollector(clazz,
-				builder);
+		IXpectURIProvider provider = new SafeURIFileCollector(clazz, builder);
 		return provider;
 	}
